North American (i.e. Canada, Mexico, and United States) student results of the 2009 Programme of International Student Assessment.

Student results from the 2009 Programme of International Student Assessment (PISA) as provided by the Organization for Economic Co-operation and Development (OECD). See http://www.pisa.oecd.org/ for more information including the code book.

Format

a data frame with 66,548 obvservations of 65 variables.

Source

Organization for Economic Co-operation and Development

Details

Note that missing values have been imputed using the http://cran.r-project.org/web/packages/mice/index.html package. Details on the specific procedure are in the pisa.impute function in the pisa package.

References

Organization for Economic Co-operation and Development (2009). Programme for International Student Assessment (PISA). http://www.pisa.oecd.org/